+++ concurrency/Control/Concurrent/Classy/Async.hs
@@ -18,14 +18,9 @@
 -- The 'withAsync' function starts an operation in a separate thread,
 -- and kills it if the inner action finishes before it completes.
 --
--- There are a few deviations from the regular async package:
---
---   * 'asyncBound' and 'withAsyncBound' are missing as @MonadConc@
---   does not support bound threads.
---
---   * The @Alternative@ instance for 'Concurrently' uses @forever
---   yield@ in the definition of @empty@, rather than @forever
---   (threadDelay maxBound)@.
+-- Unlike the regular async package, the @Alternative@ instance for
+-- 'Concurrently' uses @forever yield@ in the definition of @empty@,
+-- rather than @forever (threadDelay maxBound)@.
 module Control.Concurrent.Classy.Async
   ( -- * Asynchronous actions
     Async
@@ -33,6 +28,8 @@ module Control.Concurrent.Classy.Async
   -- * Spawning
   , async
   , asyncN
+  , asyncBound
+  , asyncBoundN
   , asyncOn
   , asyncOnN
   , asyncWithUnmask
@@ -42,6 +39,7 @@ module Control.Concurrent.Classy.Async
 
   -- * Spawning with automatic 'cancel'ation
   , withAsync
+  , withAsyncBound
   , withAsyncOn
   , withAsyncWithUnmask
   , withAsyncOnWithUnmask
@@ -188,6 +186,19 @@ async = asyncUsing fork
 asyncN :: MonadConc m => String -> m a -> m (Async m a)
 asyncN name = asyncUsing (forkN name)
 
+-- | Like 'async' but uses 'forkOS' internally.
+--
+-- @since unreleased
+asyncBound :: MonadConc m => m a -> m (Async m a)
+asyncBound = asyncUsing forkOS
+
+-- | Like 'asyncBound', but using a named thread for better debugging
+-- information.
+--
+-- @since unreleased
+asyncBoundN :: MonadConc m => String -> m a -> m (Async m a)
+asyncBoundN name = asyncUsing (forkOSN name)
+
 -- | Like 'async' but using 'forkOn' internally.
 --
 -- @since 1.1.1.0
@@ -254,6 +265,12 @@ asyncUnmaskUsing doFork action = do
 -- @since 1.1.1.0
 withAsync :: MonadConc m => m a -> (Async m a -> m b) -> m b
 withAsync = withAsyncUsing fork
+
+-- | Like 'withAsync' but uses 'forkOS' internally.
+--
+-- @since unreleased
+withAsyncBound :: MonadConc m => m a -> (Async m a -> m b) -> m b
+withAsyncBound = withAsyncUsing forkOS
 
 -- | Like 'withAsync' but uses 'forkOn' internally.
 --