+++ tasty-dejafu/Test/Tasty/DejaFu.hs
@@ -1,15 +1,7 @@
-{-# LANGUAGE CPP #-}
 {-# LANGUAGE FlexibleContexts #-}
 {-# LANGUAGE FlexibleInstances #-}
 {-# LANGUAGE GADTs #-}
-{-# LANGUAGE RankNTypes #-}
 {-# LANGUAGE TypeSynonymInstances #-}
-
-#if MIN_TOOL_VERSION_ghc(8,0,0)
--- Impredicative polymorphism checks got stronger in GHC 8, breaking
--- the use of 'unsafeCoerce' below.
-{-# LANGUAGE ImpredicativeTypes #-}
-#endif
 
 -- |
 -- Module      : Test.Tasty.DejaFu
@@ -17,23 +9,22 @@
 -- License     : MIT
 -- Maintainer  : Michael Walker <mike@barrucadu.co.uk>
 -- Stability   : stable
--- Portability : CPP, FlexibleContexts, FlexibleInstances, GADTs, ImpredicativeTypes, RankNTypes, TypeSynonymInstances
+-- Portability : FlexibleContexts, FlexibleInstances, GADTs, TypeSynonymInstances
 --
 -- This module allows using Deja Fu predicates with Tasty to test the
 -- behaviour of concurrent systems.
 module Test.Tasty.DejaFu
   ( -- * Unit testing
 
-  -- | This is supported by the 'IsTest' instances for 'ConcST' and
-  -- 'ConcIO'. These instances try all executions, reporting as
-  -- failures the cases which return a 'Just' string.
+  -- | This is supported by an 'IsTest' instance for 'ConcIO'.  This
+  -- instance tries all executions, reporting as failures the cases
+  -- which return a 'Just' string.
   --
-  -- @instance Typeable t => IsTest (ConcST t (Maybe String))@
-  -- @instance               IsTest (ConcIO   (Maybe String))@
+  -- @instance IsTest (ConcIO (Maybe String))@
   -- @instance IsOption Bounds@
   -- @instance IsOption MemType@
 
-  -- * Property testing
+  -- * Unit testing
     testAuto
   , testDejafu
   , testDejafus
@@ -43,17 +34,6 @@ module Test.Tasty.DejaFu
   , testDejafusWay
 
   , testDejafuDiscard
-
-  -- ** @IO@
-  , testAutoIO
-  , testDejafuIO
-  , testDejafusIO
-
-  , testAutoWayIO
-  , testDejafuWayIO
-  , testDejafusWayIO
-
-  , testDejafuDiscardIO
 
   -- ** Re-exports
   , Way
@@ -83,7 +63,6 @@ module Test.Tasty.DejaFu
   , R.equivalentTo, (R.===)
   ) where
 
-import           Control.Monad.ST       (runST)
 import           Data.Char              (toUpper)
 import qualified Data.Foldable          as F
 import           Data.List              (intercalate, intersperse)
@@ -101,30 +80,8 @@ import           Test.Tasty.Options     (IsOption(..), OptionDescription(..),
 import           Test.Tasty.Providers   (IsTest(..), singleTest, testFailed,
                                          testPassed)
 
--- Can't put the necessary forall in the @IsTest ConcST t@
--- instance :(
-import           Unsafe.Coerce          (unsafeCoerce)
-
-runSCTst :: (Either Failure a -> Maybe Discard) -> Way -> MemType -> (forall t. Conc.ConcST t a) -> [(Either Failure a, Conc.Trace)]
-runSCTst discard way memtype conc = runST (SCT.runSCTDiscard discard way memtype conc)
-
-runSCTio :: (Either Failure a -> Maybe Discard) -> Way -> MemType -> Conc.ConcIO a -> IO [(Either Failure a, Conc.Trace)]
-runSCTio = SCT.runSCTDiscard
-
 --------------------------------------------------------------------------------
 -- Tasty-style unit testing
-
--- | @since 0.3.0.0
-instance Typeable t => IsTest (Conc.ConcST t (Maybe String)) where
-  testOptions = Tagged concOptions
-
-  run options conc callback = do
-    let memtype = lookupOption options :: MemType
-    let way     = lookupOption options :: Way
-    let runSCTst' :: Conc.ConcST t (Maybe String) -> [(Either Failure (Maybe String), Conc.Trace)]
-        runSCTst' = unsafeCoerce $ runSCTst (const Nothing) way memtype
-    let traces = runSCTst' conc
-    run options (ConcTest traces assertableP) callback
 
 -- | @since 0.3.0.0
 instance IsTest (Conc.ConcIO (Maybe String)) where
@@ -133,8 +90,8 @@ instance IsTest (Conc.ConcIO (Maybe String)) where
   run options conc callback = do
     let memtype = lookupOption options
     let way     = lookupOption options
-    let traces  = runSCTio (const Nothing) way memtype conc
-    run options (ConcIOTest traces assertableP) callback
+    let traces  = SCT.runSCTDiscard (const Nothing) way memtype conc
+    run options (ConcTest traces assertableP) callback
 
 concOptions :: [OptionDescription]
 concOptions =
@@ -175,13 +132,9 @@ instance IsOption Way where
 -- | Automatically test a computation. In particular, look for
 -- deadlocks, uncaught exceptions, and multiple return values.
 --
--- This uses the 'Conc' monad for testing, which is an instance of
--- 'MonadConc'. If you need to test something which also uses
--- 'MonadIO', use 'testAutoIO'.
---
--- @since 0.2.0.0
+-- @since unreleased
 testAuto :: (Eq a, Show a)
-  => (forall t. Conc.ConcST t a)
+  => Conc.ConcIO a
   -- ^ The computation to test
   -> TestTree
 testAuto = testAutoWay defaultWay defaultMemType
@@ -189,30 +142,16 @@ testAuto = testAutoWay defaultWay defaultMemType
 -- | Variant of 'testAuto' which tests a computation under a given
 -- execution way and memory model.
 --
--- @since 0.5.0.0
+-- @since unreleased
 testAutoWay :: (Eq a, Show a)
   => Way
   -- ^ How to execute the concurrent program.
   -> MemType
   -- ^ The memory model to use for non-synchronised @CRef@ operations.
-  -> (forall t. Conc.ConcST t a)
+  -> Conc.ConcIO a
   -- ^ The computation to test
   -> TestTree
 testAutoWay way memtype conc = testDejafusWay way memtype conc autocheckCases
-
--- | Variant of 'testAuto' for computations which do 'IO'.
---
--- @since 0.2.0.0
-testAutoIO :: (Eq a, Show a) => Conc.ConcIO a -> TestTree
-testAutoIO = testAutoWayIO defaultWay defaultMemType
-
--- | Variant of 'testAutoWay' for computations which do 'IO'.
---
--- @since 0.5.0.0
-testAutoWayIO :: (Eq a, Show a)
-  => Way -> MemType -> Conc.ConcIO a -> TestTree
-testAutoWayIO way memtype concio =
-  testDejafusWayIO way memtype  concio autocheckCases
 
 -- | Predicates for the various autocheck functions.
 autocheckCases :: Eq a => [(TestName, Predicate a)]
@@ -224,9 +163,9 @@ autocheckCases =
 
 -- | Check that a predicate holds.
 --
--- @since 0.2.0.0
+-- @since unreleased
 testDejafu :: Show a
-  => (forall t. Conc.ConcST t a)
+  => Conc.ConcIO a
   -- ^ The computation to test
   -> TestName
   -- ^ The name of the test.
@@ -238,13 +177,13 @@ testDejafu = testDejafuWay defaultWay defaultMemType
 -- | Variant of 'testDejafu' which takes a way to execute the program
 -- and a memory model.
 --
--- @since 0.5.0.0
+-- @since unreleased
 testDejafuWay :: Show a
   => Way
   -- ^ How to execute the concurrent program.
   -> MemType
   -- ^ The memory model to use for non-synchronised @CRef@ operations.
-  -> (forall t. Conc.ConcST t a)
+  -> Conc.ConcIO a
   -- ^ The computation to test
   -> TestName
   -- ^ The name of the test.
@@ -255,7 +194,7 @@ testDejafuWay = testDejafuDiscard (const Nothing)
 
 -- | Variant of 'testDejafuWay' which can selectively discard results.
 --
--- @since 0.7.0.0
+-- @since unreleased
 testDejafuDiscard :: Show a
   => (Either Failure a -> Maybe Discard)
   -- ^ Selectively discard results.
@@ -263,7 +202,7 @@ testDejafuDiscard :: Show a
   -- ^ How to execute the concurrent program.
   -> MemType
   -- ^ The memory model to use for non-synchronised @CRef@ operations.
-  -> (forall t. Conc.ConcST t a)
+  -> Conc.ConcIO a
   -- ^ The computation to test
   -> String
   -- ^ The name of the test.
@@ -271,15 +210,15 @@ testDejafuDiscard :: Show a
   -- ^ The predicate to check
   -> TestTree
 testDejafuDiscard discard way memtype conc name test =
-  testst discard way memtype conc [(name, test)]
+  testconc discard way memtype conc [(name, test)]
 
 -- | Variant of 'testDejafu' which takes a collection of predicates to
 -- test. This will share work between the predicates, rather than
 -- running the concurrent computation many times for each predicate.
 --
--- @since 0.2.0.0
+-- @since unreleased
 testDejafus :: Show a
-  => (forall t. Conc.ConcST t a)
+  => Conc.ConcIO a
   -- ^ The computation to test
   -> [(TestName, Predicate a)]
   -- ^ The list of predicates (with names) to check
@@ -289,51 +228,18 @@ testDejafus = testDejafusWay defaultWay defaultMemType
 -- | Variant of 'testDejafus' which takes a way to execute the program
 -- and a memory model.
 --
--- @since 0.5.0.0
+-- @since unreleased
 testDejafusWay :: Show a
   => Way
   -- ^ How to execute the concurrent program.
   -> MemType
   -- ^ The memory model to use for non-synchronised @CRef@ operations.
-  -> (forall t. Conc.ConcST t a)
+  -> Conc.ConcIO a
   -- ^ The computation to test
   -> [(TestName, Predicate a)]
   -- ^ The list of predicates (with names) to check
   -> TestTree
-testDejafusWay = testst (const Nothing)
-
--- | Variant of 'testDejafu' for computations which do 'IO'.
---
--- @since 0.2.0.0
-testDejafuIO :: Show a => Conc.ConcIO a -> TestName -> Predicate a -> TestTree
-testDejafuIO = testDejafuWayIO defaultWay defaultMemType
-
--- | Variant of 'testDejafuWay' for computations which do 'IO'.
---
--- @since 0.5.0.0
-testDejafuWayIO :: Show a
-  => Way -> MemType -> Conc.ConcIO a -> TestName -> Predicate a -> TestTree
-testDejafuWayIO = testDejafuDiscardIO (const Nothing)
-
--- | Variant of 'testDejafuDiscard' for computations which do 'IO'.
---
--- @since 0.7.0.0
-testDejafuDiscardIO :: Show a => (Either Failure a -> Maybe Discard) -> Way -> MemType -> Conc.ConcIO a -> String -> Predicate a -> TestTree
-testDejafuDiscardIO discard way memtype concio name test =
-  testio discard way memtype concio [(name, test)]
-
--- | Variant of 'testDejafus' for computations which do 'IO'.
---
--- @since 0.2.0.0
-testDejafusIO :: Show a => Conc.ConcIO a -> [(TestName, Predicate a)] -> TestTree
-testDejafusIO = testDejafusWayIO defaultWay defaultMemType
-
--- | Variant of 'dejafusWay' for computations which do 'IO'.
---
--- @since 0.5.0.0
-testDejafusWayIO :: Show a
-  => Way -> MemType -> Conc.ConcIO a -> [(TestName, Predicate a)] -> TestTree
-testDejafusWayIO = testio (const Nothing)
+testDejafusWay = testconc (const Nothing)
 
 
 -------------------------------------------------------------------------------
@@ -356,11 +262,7 @@ testProperty = testprop
 -- Tasty integration
 
 data ConcTest where
-  ConcTest   :: Show a => [(Either Failure a, Conc.Trace)] -> Predicate a -> ConcTest
-  deriving Typeable
-
-data ConcIOTest where
-  ConcIOTest :: Show a => IO [(Either Failure a, Conc.Trace)] -> Predicate a -> ConcIOTest
+  ConcTest :: Show a => IO [(Either Failure a, Conc.Trace)] -> Predicate a -> ConcTest
   deriving Typeable
 
 data PropTest where
@@ -370,14 +272,7 @@ data PropTest where
 instance IsTest ConcTest where
   testOptions = pure []
 
-  run _ (ConcTest traces p) _ =
-    let err = showErr $ p traces
-     in pure (if null err then testPassed "" else testFailed err)
-
-instance IsTest ConcIOTest where
-  testOptions = pure []
-
-  run _ (ConcIOTest iotraces p) _ = do
+  run _ (ConcTest iotraces p) _ = do
     traces <- iotraces
     let err = showErr $ p traces
     pure (if null err then testPassed "" else testFailed err)
@@ -397,41 +292,24 @@ instance IsTest PropTest where
         ]
       Nothing -> testPassed ""
 
--- | Produce a Tasty 'TestTree' from a Deja Fu test.
-testst :: Show a
-  => (Either Failure a -> Maybe Discard)
-  -> Way
-  -> MemType
-  -> (forall t. Conc.ConcST t a)
-  -> [(TestName, Predicate a)]
-  -> TestTree
-testst discard way memtype conc tests = case map toTest tests of
-  [t] -> t
-  ts  -> testGroup "Deja Fu Tests" ts
-
-  where
-    toTest (name, p) = singleTest name $ ConcTest traces p
-
-    traces = runSCTst discard way memtype conc
-
--- | Produce a Tasty 'Test' from an IO-using Deja Fu test.
-testio :: Show a
+-- | Produce a Tasty 'Test' from a Deja Fu unit test.
+testconc :: Show a
   => (Either Failure a -> Maybe Discard)
   -> Way
   -> MemType
   -> Conc.ConcIO a
   -> [(TestName, Predicate a)]
   -> TestTree
-testio discard way memtype concio tests = case map toTest tests of
+testconc discard way memtype concio tests = case map toTest tests of
   [t] -> t
   ts  -> testGroup "Deja Fu Tests" ts
 
   where
-    toTest (name, p) = singleTest name $ ConcIOTest traces p
+    toTest (name, p) = singleTest name $ ConcTest traces p
 
     -- As with HUnit, constructing a test is side-effect free, so
     -- sharing of traces can't happen here.
-    traces = runSCTio discard way memtype concio
+    traces = SCT.runSCTDiscard discard way memtype concio
 
 -- | Produce a Tasty 'TestTree' from a Deja Fu refinement property test.
 testprop :: (R.Testable p, R.Listable (R.X p), Eq (R.X p), Show (R.X p), Show (R.O p))