+++ concurrency/CHANGELOG.markdown
@@ -7,6 +7,30 @@ This project is versioned according to the [Package Versioning Policy](https://p
 *de facto* standard Haskell versioning scheme.
 
 
+unreleased
+----------
+
+### Control.Concurrent.Classy.Async
+
+- New `asyncBound`, `asyncBoundN`, and `withAsyncBound` functions for doing asynchronous actions on
+  bound threads. (#126)
+
+### Control.Monad.Conc.Class
+
+- `MonadConc` now supports bound threads with new `forkOS`, `forkOSN`, and `isCurrentThreadBound`
+  functions. (#126)
+
+- New `runInBoundThread` and `runInUnboundThread` functions. (#126)
+
+- The `rtsSupportsBoundThreads` definition is now the definition from Control.Concurrent
+  re-exported, not just `False`. (#126)
+
+Note that bound threads are only supported if you compile with GHC and link with -threaded.
+
+
+---------------------------------------------------------------------------------------------------
+
+
 1.2.2.0
 -------